Learn & Implement MongoDB Aggregation Pipelines
There is a need to use aggregation pipelines in this project, but i don't know what this concept was before, so i took all the days it took me to learn & implement "aggregation pipelines" and i made notes & documentation in this file about all the things i learned about "aggregation pipelines".
I am not a expert at "aggregation pipelines" but i can handle most of the scenarios now. Please open this markdown file to see what i learned and implemented.